timestamp_utils/tokenizer_utils/types re-organization
