+++ examples/models/parakeet/CMakeLists.txt
@@ -80,7 +80,12 @@ if(EXECUTORCH_BUILD_METAL)
   executorch_target_link_options_shared_lib(metal_backend)
 endif()
 
-add_executable(parakeet_runner main.cpp)
+add_executable(
+  parakeet_runner
+  main.cpp
+  timestamp_utils.cpp
+  tokenizer_utils.cpp
+)
 if(NOT CMAKE_BUILD_TYPE STREQUAL "Debug")
   target_link_options_gc_sections(parakeet_runner)
   if(NOT APPLE AND NOT MSVC)
